backend/pyproject.toml
[project]
name = "paper-boy-backend"
version = "0.1.0"
description = "AI-powered custom newsletter delivery service"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.3",
    "django-environ>=0.11",
    "psycopg[binary]>=3.1",
    "celery[redis]>=5.3",
    "django-celery-beat>=2.6",
    "anthropic>=0.28",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "dj-email-url>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "factory-boy>=3.3",
    "ruff>=0.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "factory-boy>=3.3",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py", "*_test.py"]
